+++ test/harness/jpype/properties/TestBean.java
@@ -9,6 +9,12 @@ public class TestBean {
 
 	private String property4;
 
+	private String property5;
+
+	private String property6;
+
+	private String property7;
+
 	public String getProperty1() {
 		return "get" + property1;
 	}
@@ -25,6 +31,10 @@ public class TestBean {
 		return "abc" + property4;
 	}
 
+	public String getProperty6() {
+		return "get" + property7;
+	}
+
 	public String property1() {
 		return "method";
 	}
@@ -37,6 +47,10 @@ public class TestBean {
 		return "method";
 	}
 
+	public String returnProperty5() {
+		return "return" + this.property5;
+	}
+
 	public void setProperty1(String property1) {
 		this.property1 = "set" + property1;
 	}
@@ -47,5 +61,13 @@ public class TestBean {
 
 	public void setProperty3(String property3) {
 		this.property3 = "set" + property3;
+	}
+
+	public void setProperty5(String property5) {
+		this.property5 = "set" + property5;
+	}
+	
+	public void setProperty6(String property6) {
+		this.property7 = "set" + property6;
 	}
 }